[api] Add roomCreation api as well as memberships
The api can now create rooms on a matrix server and invite people to it.
It also includes the very similar methods: kick, leave, forget, ban,
unban that can be use for moderation inside a room.

The joinedRooms endpoint now return a list of RoomId instead of
returning a list of string.