chore(refactor): introduce JsonRequest and update MatrixClient  (#87)
The primary goal of this change was to remove the dependency on MatrixConfiguration when creating a request using the `get`, `post`, `postJson` and other helpers. This also made possible to remove the pretty confusing type alias of `MatrixAction` and delayed the creation of the actual request up until the client itself.

The `JsonRequest` class can the create a Sttp request based on the configuration. As of now it only support a subset of configuration (required to make this library work in the current state):

 - Authentication
 - Query parameters
 - Json and Empty bodies
 
 This also assumes that the API will always return a JSON, it might not be correct but I could not find a counter example in the documentation.
In more customization is needed in the future, this class could wrap a `RequestT` from Sttp to allow infinite customization.