chore(refactor): merge endpoints inside Matrix (#88)
Those changes are based on the new API used in the zio-slack. This aims to respect a bit more the ZIO guideline to expose methods. This is also easier to discover which methods are available in the API as they are not hidden behind different traits.

All methods are now exposed inside a single `Matrix` object, but the implementation is splitted into multiple traits and accessors.
The `MatrixRequest` now only contains helpers to build `JsonRequest` and the api package object has been as methods have been moved inside the `MatrixApiBase` trait.

This is not perfect yet as this trait contains a lot of information and does not provide a clear interface that could be used to mock/replace/test, but is has the benefit of being able to use it as a mixin inside the multiple API traits.